feat: store real region in region_config
