+++ lib/region_config.js
@@ -1,13 +1,8 @@
 var util = require('./util');
 var regionConfig = require('./region_config_data.json');
 
-function generateRegionPrefix(region, options) {
+function generateRegionPrefix(region) {
   if (!region) return null;
-
-  if (options.useFipsEndpoint) {
-    region = getRealRegion(region);
-  }
-
   var parts = region.split('-');
   if (parts.length < 3) return null;
   return parts.slice(0, parts.length - 2).join('-') + '-*';
@@ -15,12 +10,7 @@ function generateRegionPrefix(region, options) {
 
 function derivedKeys(service) {
   var region = service.config.region;
-
-  var useFipsEndpoint = service.config.useFipsEndpoint;
-  var regionPrefix = generateRegionPrefix(region, {
-    useFipsEndpoint: useFipsEndpoint
-  });
-
+  var regionPrefix = generateRegionPrefix(region);
   var endpointPrefix = service.api.endpointPrefix;
 
   return [
@@ -101,25 +91,10 @@ function getEndpointSuffix(region) {
   return defaultSuffix;
 }
 
-function isFipsRegion(region) {
-  return region && (region.startsWith('fips-') || region.endsWith('-fips'));
-}
-
-function getRealRegion(region) {
-  return isFipsRegion(region)
-    ? ['fips-aws-global', 'aws-fips'].includes(region)
-      ? 'us-east-1'
-      : region === 'fips-aws-us-gov-global'
-      ? 'us-gov-west-1'
-      : region.replace(/fips-(dkr-|prod-)?|-fips/, '')
-    : region;
-}
-
 /**
  * @api private
  */
 module.exports = {
   configureEndpoint: configureEndpoint,
   getEndpointSuffix: getEndpointSuffix,
-  getRealRegion: getRealRegion,
 };
