+++ lib/region_config_data.json
@@ -105,12 +105,12 @@
     "us-gov-*/ssm": "fipsWithServiceOnly",
     "us-gov-*/sts": "fipsWithServiceOnly",
     "us-gov-*/support": "fipsWithServiceOnly",
-    "fips-us-gov-west-1/states": "fipsWithServiceOnly",
-    "fips-us-iso-east-1/elasticfilesystem": {
+    "us-gov-west-1/states": "fipsWithServiceOnly",
+    "us-iso-east-1/elasticfilesystem": {
       "endpoint": "elasticfilesystem-fips.{region}.c2s.ic.gov"
     },
-    "fips-aws-us-gov-global/organizations": "fipsWithServiceOnly",
-    "fips-aws-us-gov-global/route53": {
+    "us-gov-west-1/organizations": "fipsWithServiceOnly",
+    "us-gov-west-1/route53": {
       "endpoint": "route53.us-gov.amazonaws.com"
     }
   },
