+++ lib/service.js
@@ -4,6 +4,7 @@ var regionConfig = require('./region_config');
 
 var inherit = AWS.util.inherit;
 var clientCount = 0;
+var region_utils = require('./region/utils');
 
 /**
  * The service class representing an AWS service.
@@ -25,6 +26,13 @@ AWS.Service = inherit({
       throw AWS.util.error(new Error(),
         'Service must be constructed with `new\' operator');
     }
+
+    var region = config.region;
+    if (region && region_utils.isFipsRegion(region)) {
+      config.region = region_utils.getRealRegion(region);
+      config.useFipsEndpoint = true;
+    }
+
     var ServiceClass = this.loadServiceClass(config || {});
     if (ServiceClass) {
       var originalConfig = AWS.util.copy(config);
@@ -636,7 +644,7 @@ AWS.Service = inherit({
 
     var e = endpoint;
     e = e.replace(/\{service\}/g, this.api.endpointPrefix);
-    e = e.replace(/\{region\}/g, regionConfig.getRealRegion(this.config.region));
+    e = e.replace(/\{region\}/g, this.config.region);
     e = e.replace(/\{scheme\}/g, this.config.sslEnabled ? 'https' : 'http');
     return e;
   },
